feat: add gen and pri domain validation lambdas
