--- src/serverlessfns/hello/index.ts
@@ -1,15 +1,0 @@
-import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda"
-
-export const hello = async (
-  event: APIGatewayProxyEvent
-): Promise<APIGatewayProxyResult> => ({
-  statusCode: 200,
-  body: JSON.stringify(
-    {
-      message: "Go Serverless v3.0! Your function executed successfully!",
-      input: event,
-    },
-    null,
-    2
-  ),
-})
