--- src/serverlessfns/step-functions/first-step-function.asl.yaml
@@ -1,11 +1,0 @@
-Comment: "Two consecutive lambdas"
-StartAt: Hello
-States:
-  Hello:
-    Type: Task
-    Resource: "arn:aws:lambda:ap-southeast-1:101010101010:function:${self:service}-${opt:stage, 'dev'}-hello"
-    Next: HiThere
-  HiThere:
-    Type: Task
-    Resource: "arn:aws:lambda:ap-southeast-1:101010101010:function:${self:service}-${opt:stage, 'dev'}-hithere"
-    End: true
